Final Project
This project was supposed to present the skills I acquired while participating in class and doing assignments. Additionally, this is a group project, but I have excluded the information of other members. This report was supposed to reflect on students with Depression and classify the data based on the features in the data set. Although this dataset was obtained through an online open-source platform with a usability of 9.41, we needed to discretize the data to optimize the efficiency of fitting the models. Models used are KNN, Naive Bayes, Random Forest, ANN, CART, and SVM.